+++ springboot.jpa.mysql/src/main/java/com/akash/springboot/jpa/mysql/dao/UserDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.akash.springboot.jpa.mysql.entity.User;
 
-public interface UserDao extends JpaRepository<User, Long> {
+public interface UserDao extends JpaRepository<User, Integer> {
 
 }
